Fixed and tested the gateway, implemented work with the user service and media service via grpc

+++ user/internal/storage/postgresql/friends.go
@@ -2,9 +2,9 @@ package postgresql
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/jackc/pgconn"
+	"github.com/jackc/pgerrcode"
 	"gomessage.com/users/internal/models"
 	"gomessage.com/users/internal/storage"
 	"gomessage.com/users/pkg/postgresql"
@@ -14,8 +14,10 @@ type friendsRepository struct {
 	client postgresql.Client
 }
 
+const nil_string = " "
+
 // Create implements storage.FriendsRepository.
-func (f *friendsRepository) Create(ctx context.Context, friend *models.FriendListModel) error {
+func (f *friendsRepository) Create(ctx context.Context, friend *models.FriendListModel) (string, error) {
 	q := `INSERT INTO friend_list
 			(user_id, friend_id) 
 		VALUES 
@@ -27,17 +29,22 @@ func (f *friendsRepository) Create(ctx context.Context, friend *models.FriendLis
 	err := row.Scan(&friend.ID)
 	if err != nil {
 		if pgErr, ok := err.(*pgconn.PgError); ok {
-			fmt.Println(fmt.Sprintf("Sql Error: %s, Detail %s, Where %s", pgErr.Message, pgErr.Detail, pgErr.Where))
-			return err
+			var info_err string
+			if pgErr.Code == pgerrcode.UniqueViolation {
+				info_err = "You already friends"
+			} else if pgErr.Code == pgerrcode.InvalidTextRepresentation {
+				info_err = "Invalid friend id"
+			}
+			return info_err, err
 		}
-		return err
+		return nil_string, err
 	}
-	return nil
+	return nil_string, nil
 }
 
 func (f *friendsRepository) FindAll(ctx context.Context, user_id string) ([]models.FriendListModel, error) {
-	q := `SELECT * FROM friend_list WHERE user_id=$1`
-	qrow, err := f.client.Query(ctx, q)
+	q := `SELECT friend_id FROM friend_list WHERE user_id=$1`
+	qrow, err := f.client.Query(ctx, q, user_id)
 	if err != nil {
 		return nil, err
 	}
@@ -45,7 +52,7 @@ func (f *friendsRepository) FindAll(ctx context.Context, user_id string) ([]mode
 	friends := make([]models.FriendListModel, 0)
 	for qrow.Next() {
 		var frn models.FriendListModel
-		err := qrow.Scan(&frn.ID, &frn.UserID, &frn.FriendID)
+		err := qrow.Scan(&frn.FriendID)
 		if err != nil {
 			return nil, err
 		}
@@ -59,9 +66,11 @@ func (f *friendsRepository) FindAll(ctx context.Context, user_id string) ([]mode
 
 // Delete implements storage.FriendsRepository.
 func (f *friendsRepository) Delete(ctx context.Context, friends *models.FriendListModel) error {
-	query := `DELETE FROM friend_list WHERE user_id = $1 AND friend_id = $2`
-	f.client.QueryRow(ctx, query, friends.UserID, friends.FriendID)
-
+	query := `DELETE FROM friend_list WHERE user_id = $1 AND friend_id = $2 OR friend_id=$1 AND user_id =$2`
+	_, err := f.client.Exec(ctx, query, friends.UserID, friends.FriendID)
+	if err != nil {
+		return err
+	}
 	return nil
 }
